add branch route
