--- public/javascripts/restaurantView.js
@@ -1,5 +1,0 @@
-const branchController = require('../../server/controllers/branch');
-
-function check() {
-    alert();
-}
