+++ server/controllers/home.js
@@ -1,7 +1,8 @@
 'use strict';
 const db = require('../../server/helpers/database').db;
 const searchQuery = require('../../sqlQueries/searchFoodItems');
-const restaurantsQuery = require('../../sqlQueries/restaurants');
+const restaurantController = require('../controllers/restaurant');
+const branchController = require('../controllers/branch');
 const passport = require('passport');
 
 let index = (req, res) => {
@@ -86,31 +87,8 @@ let search = (req, res) => {
     })(req, res, next);
 };
 
-let viewRestaurants = (req, res) => {
-    let dictionary = [];
-    console.log('Retrieving all restaurants');
-    db.query(restaurantsQuery.allRestaurantsAndBranches)
-        .then(val => {
-            if (!val || val.rowCount == 0) return;
-            let rows = val.rows;
+let viewRestaurants = restaurantController(req, res);
 
-            for (let i = 0; i < val.rowCount; i++) {
-                let rid = rows[i].restaurant_id - 1;
-                let branch = { bid: rows[i].branches_id, bname: rows[i].bname, bphone: rows[i].bphone, baddress: rows[i].baddress, barea: rows[i].barea };
-                if (dictionary[rid] == undefined) {
-                    let branches = [];
-                    branches.push(branch);
-                    dictionary[rid] = { rid: rows[i].restaurant_id, rname: rows[i].rname, rphone: rows[i].rphone, raddress: rows[i].raddress, branches: branches };
+let getBranch = branchController.getBranch(req, res);
 
-                } else {
-                    let branches = dictionary[rid].branches;
-                    branches.push(branch);
-                    dictionary[rid].branches = branches;
-                }
-            }
-            console.log(JSON.stringify(dictionary));
-            res.render('restaurants', { layout: 'index', title: 'All Restaurants', restaurants: dictionary });
-        });
-};
-
-module.exports = { index: index, handleLoginValidation: handleLoginValidation, search: search, viewRestaurants: viewRestaurants };
+module.exports = { index: index, handleLoginValidation: handleLoginValidation, search: search, viewRestaurants: viewRestaurants, getBranch: getBranch };
