+++ server/routes/home.js
@@ -8,5 +8,6 @@ router.get('/', homeController.index);
 router.post('/', homeController.handleLoginValidation);
 router.post('/search', homeController.search);
 router.get('/restaurants', homeController.viewRestaurants);
+router.get('/branch/:bid', homeController.getBranches);
 
 module.exports = router;
